+++ src/object_pool/object_type.rs
@@ -1,0 +1,170 @@
+use crate::object_pool::ParseError;
+use crate::object_pool::ParseError::UnknownObjectType;
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum ObjectType {
+    WorkingSet = 0,
+    DataMask = 1,
+    AlarmMask = 2,
+    Container = 3,
+    SoftKeyMask = 4,
+    Key = 5,
+    Button = 6,
+    InputBoolean = 7,
+    InputString = 8,
+    InputNumber = 9,
+    InputList = 10,
+    OutputString = 11,
+    OutputNumber = 12,
+    OutputLine = 13,
+    OutputRectangle = 14,
+    OutputEllipse = 15,
+    OutputPolygon = 16,
+    OutputMeter = 17,
+    OutputLinearBarGraph = 18,
+    OutputArchedBarGraph = 19,
+    PictureGraphic = 20,
+    NumberVariable = 21,
+    StringVariable = 22,
+    FontAttributes = 23,
+    LineAttributes = 24,
+    FillAttributes = 25,
+    InputAttributes = 26,
+    ObjectPointer = 27,
+    Macro = 28,
+    AuxiliaryFunctionType1 = 29,
+    AuxiliaryInputType1 = 30,
+    AuxiliaryFunctionType2 = 31,
+    AuxiliaryInputType2 = 32,
+    AuxiliaryControlDesignatorType2 = 33,
+    WindowMask = 34,
+    KeyGroup = 35,
+    GraphicsContext = 36,
+    OutputList = 37,
+    ExtendedInputAttributes = 38,
+    ColourMap = 39,
+    ObjectLabelReferenceList = 40,
+    ExternalObjectDefinition = 41,
+    ExternalReferenceName = 42,
+    ExternalObjectPointer = 43,
+    Animation = 44,
+    ColourPalette = 45,
+    GraphicData = 46,
+    WorkingSetSpecialControls = 47,
+    ScaledGraphic = 48,
+}
+
+impl TryFrom<u8> for ObjectType {
+    type Error = ParseError;
+
+    fn try_from(val: u8) -> Result<Self, Self::Error> {
+        match val {
+            0 => Ok(Self::WorkingSet),
+            1 => Ok(Self::DataMask),
+            2 => Ok(Self::AlarmMask),
+            3 => Ok(Self::Container),
+            4 => Ok(Self::SoftKeyMask),
+            5 => Ok(Self::Key),
+            6 => Ok(Self::Button),
+            7 => Ok(Self::InputBoolean),
+            8 => Ok(Self::InputString),
+            9 => Ok(Self::InputNumber),
+            10 => Ok(Self::InputList),
+            11 => Ok(Self::OutputString),
+            12 => Ok(Self::OutputNumber),
+            13 => Ok(Self::OutputLine),
+            14 => Ok(Self::OutputRectangle),
+            15 => Ok(Self::OutputEllipse),
+            16 => Ok(Self::OutputPolygon),
+            17 => Ok(Self::OutputMeter),
+            18 => Ok(Self::OutputLinearBarGraph),
+            19 => Ok(Self::OutputArchedBarGraph),
+            20 => Ok(Self::PictureGraphic),
+            21 => Ok(Self::NumberVariable),
+            22 => Ok(Self::StringVariable),
+            23 => Ok(Self::FontAttributes),
+            24 => Ok(Self::LineAttributes),
+            25 => Ok(Self::FillAttributes),
+            26 => Ok(Self::InputAttributes),
+            27 => Ok(Self::ObjectPointer),
+            28 => Ok(Self::Macro),
+            29 => Ok(Self::AuxiliaryFunctionType1),
+            30 => Ok(Self::AuxiliaryInputType1),
+            31 => Ok(Self::AuxiliaryFunctionType2),
+            32 => Ok(Self::AuxiliaryInputType2),
+            33 => Ok(Self::AuxiliaryControlDesignatorType2),
+            34 => Ok(Self::WindowMask),
+            35 => Ok(Self::KeyGroup),
+            36 => Ok(Self::GraphicsContext),
+            37 => Ok(Self::OutputList),
+            38 => Ok(Self::ExtendedInputAttributes),
+            39 => Ok(Self::ColourMap),
+            40 => Ok(Self::ObjectLabelReferenceList),
+            41 => Ok(Self::ExternalObjectDefinition),
+            42 => Ok(Self::ExternalReferenceName),
+            43 => Ok(Self::ExternalObjectPointer),
+            44 => Ok(Self::Animation),
+            45 => Ok(Self::ColourPalette),
+            46 => Ok(Self::GraphicData),
+            47 => Ok(Self::WorkingSetSpecialControls),
+            48 => Ok(Self::ScaledGraphic),
+            _ => Err(UnknownObjectType),
+        }
+    }
+}
+
+impl From<ObjectType> for u8 {
+    fn from(val: ObjectType) -> Self {
+        match val {
+            ObjectType::WorkingSet => 0,
+            ObjectType::DataMask => 1,
+            ObjectType::AlarmMask => 2,
+            ObjectType::Container => 3,
+            ObjectType::SoftKeyMask => 4,
+            ObjectType::Key => 5,
+            ObjectType::Button => 6,
+            ObjectType::InputBoolean => 7,
+            ObjectType::InputString => 8,
+            ObjectType::InputNumber => 9,
+            ObjectType::InputList => 10,
+            ObjectType::OutputString => 11,
+            ObjectType::OutputNumber => 12,
+            ObjectType::OutputLine => 13,
+            ObjectType::OutputRectangle => 14,
+            ObjectType::OutputEllipse => 15,
+            ObjectType::OutputPolygon => 16,
+            ObjectType::OutputMeter => 17,
+            ObjectType::OutputLinearBarGraph => 18,
+            ObjectType::OutputArchedBarGraph => 19,
+            ObjectType::PictureGraphic => 20,
+            ObjectType::NumberVariable => 21,
+            ObjectType::StringVariable => 22,
+            ObjectType::FontAttributes => 23,
+            ObjectType::LineAttributes => 24,
+            ObjectType::FillAttributes => 25,
+            ObjectType::InputAttributes => 26,
+            ObjectType::ObjectPointer => 27,
+            ObjectType::Macro => 28,
+            ObjectType::AuxiliaryFunctionType1 => 29,
+            ObjectType::AuxiliaryInputType1 => 30,
+            ObjectType::AuxiliaryFunctionType2 => 31,
+            ObjectType::AuxiliaryInputType2 => 32,
+            ObjectType::AuxiliaryControlDesignatorType2 => 33,
+            ObjectType::WindowMask => 34,
+            ObjectType::KeyGroup => 35,
+            ObjectType::GraphicsContext => 36,
+            ObjectType::OutputList => 37,
+            ObjectType::ExtendedInputAttributes => 38,
+            ObjectType::ColourMap => 39,
+            ObjectType::ObjectLabelReferenceList => 40,
+            ObjectType::ExternalObjectDefinition => 41,
+            ObjectType::ExternalReferenceName => 42,
+            ObjectType::ExternalObjectPointer => 43,
+            ObjectType::Animation => 44,
+            ObjectType::ColourPalette => 45,
+            ObjectType::GraphicData => 46,
+            ObjectType::WorkingSetSpecialControls => 47,
+            ObjectType::ScaledGraphic => 48,
+        }
+    }
+}
