+++ src/object_pool/writer.rs
@@ -1,5 +1,19 @@
 use super::*;
 use crate::object_pool::colour::Colour;
+use crate::object_pool::object::{
+    AlarmMask, Animation, AuxiliaryControlDesignatorType2, AuxiliaryFunctionType1,
+    AuxiliaryFunctionType2, AuxiliaryInputType1, AuxiliaryInputType2, Button, CharacterRange,
+    CodePlane, ColourMap, ColourPalette, Container, DataMask, ExtendedInputAttributes,
+    ExternalObjectDefinition, ExternalObjectPointer, ExternalReferenceName, FillAttributes,
+    FontAttributes, GraphicData, GraphicsContext, InputAttributes, InputBoolean, InputList,
+    InputNumber, InputString, Key, KeyGroup, LineAttributes, Macro, NumberVariable, Object,
+    ObjectLabelReferenceList, ObjectPointer, OutputArchedBarGraph, OutputEllipse, OutputLine,
+    OutputLinearBarGraph, OutputList, OutputMeter, OutputNumber, OutputPolygon, OutputRectangle,
+    OutputString, PictureGraphic, ScaledGraphic, SoftKeyMask, StringVariable, WindowMask,
+    WorkingSet, WorkingSetSpecialControls,
+};
+use crate::object_pool::object_attributes::{MacroRef, ObjectLabel, ObjectRef, Point};
+use crate::object_pool::object_id::ObjectId;
 
 impl Object {
     pub fn write(&self) -> Vec<u8> {
